requirements/base.txt
django==3.2.0
django-safedelete==1.2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
